Changed naming conventions accorfing to new naming conventions of Carbon-Messaging. Began to add new sample for chat app using WebSocket support.
