+++ core/src/main/java/org/wso2/msf4j/internal/websocket/SessionManager.java
@@ -20,7 +20,7 @@ package org.wso2.msf4j.internal.websocket;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.wso2.carbon.messaging.websocket.WebSocketCarbonMessage;
+import org.wso2.carbon.messaging.websocket.WebSocketMessage;
 import org.wso2.carbon.messaging.websocket.WebSocketResponder;
 import org.wso2.carbon.transport.http.netty.common.Constants;
 
@@ -48,45 +48,45 @@ public class SessionManager {
     }
 
     /**
-     * @param webSocketCarbonMessage incoming {@link WebSocketCarbonMessage}
+     * @param webSocketMessage incoming {@link WebSocketMessage}
      * @return requested {@link Session} for given channel
      */
-    public Session getSession(WebSocketCarbonMessage webSocketCarbonMessage) {
-        String sessionId = getSessionId(webSocketCarbonMessage);
+    public Session getSession(WebSocketMessage webSocketMessage) {
+        String sessionId = getSessionId(webSocketMessage);
         if (sessionMap.containsKey(sessionId)) {
             return sessionMap.get(sessionId);
         } else {
-            return createSession(webSocketCarbonMessage);
+            return createSession(webSocketMessage);
         }
     }
 
     /**
      * This method creates session for a given channel
      * Here unlike http channel ID was taken as the session ID
-     * @param webSocketCarbonMessage incoming {@link WebSocketCarbonMessage}
+     * @param webSocketMessage incoming {@link WebSocketMessage}
      * @return created {@link Session} for given channel
      */
-    public Session createSession(WebSocketCarbonMessage webSocketCarbonMessage) {
-        String sessionId = getSessionId(webSocketCarbonMessage);
-        WebSocketResponder webSocketResponder = webSocketCarbonMessage.getWebSocketResponder();
-        URI uri = (URI) webSocketCarbonMessage.getProperty(Constants.TO);
+    public Session createSession(WebSocketMessage webSocketMessage) {
+        String sessionId = getSessionId(webSocketMessage);
+        WebSocketResponder webSocketResponder = webSocketMessage.getWebSocketResponder();
+        URI uri = (URI) webSocketMessage.getProperty(Constants.TO);
         SessionImpl session = new SessionImpl(webSocketResponder, sessionId, uri, true, true, null, null, null);
         sessionMap.put(sessionId, session);
         log.info("Session created for channel " + sessionId);
         return session;
     }
 
-    public boolean containsSession(WebSocketCarbonMessage webSocketCarbonMessage) {
-        String sessionId = getSessionId(webSocketCarbonMessage);
+    public boolean containsSession(WebSocketMessage webSocketMessage) {
+        String sessionId = getSessionId(webSocketMessage);
         return sessionMap.containsKey(sessionId);
     }
 
     /**
      * Close the channel for given session and remove session from the session manager
-     * @param webSocketCarbonMessage
+     * @param webSocketMessage
      */
-    public void removeSession(WebSocketCarbonMessage webSocketCarbonMessage) throws IOException {
-        String sessionId = getSessionId(webSocketCarbonMessage);
+    public void removeSession(WebSocketMessage webSocketMessage) throws IOException {
+        String sessionId = getSessionId(webSocketMessage);
         if (sessionMap.containsKey(sessionId)) {
             sessionMap.remove(sessionId).close();
             log.info("Removed session ID for channel " + sessionId);
@@ -96,8 +96,8 @@ public class SessionManager {
     }
 
 
-    private String getSessionId(WebSocketCarbonMessage webSocketCarbonMessage) {
-        return (String) webSocketCarbonMessage.getProperty(Constants.CHANNEL_ID);
+    private String getSessionId(WebSocketMessage webSocketMessage) {
+        return (String) webSocketMessage.getProperty(Constants.CHANNEL_ID);
     }
