+++ core/src/test/java/org/wso2/msf4j/websocket/TestEndpoint.java
@@ -1,0 +1,53 @@
+/*
+ *   Copyright (c) ${date}, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ *   WSO2 Inc. licenses this file to you under the Apache License,
+ *   Version 2.0 (the "License"); you may not use this file except
+ *   in compliance with the License.
+ *   You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing,
+ *  software distributed under the License is distributed on an
+ *  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ *  KIND, either express or implied.  See the License for the
+ *  specific language governing permissions and limitations
+ *  under the License.
+ */
+
+package org.wso2.msf4j.websocket;
+
+import org.osgi.service.component.annotations.Component;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.wso2.msf4j.WebSocketEndpoint;
+
+import javax.websocket.OnClose;
+import javax.websocket.OnMessage;
+import javax.websocket.OnOpen;
+import javax.websocket.Session;
+import javax.websocket.server.ServerEndpoint;
+
+@Component(
+
+)
+@ServerEndpoint("/test")
+public class TestEndpoint implements WebSocketEndpoint {
+    Logger log = LoggerFactory.getLogger(TestEndpoint.class);
+
+    @OnOpen
+    public void onOpen(Session session) {
+        log.info(session.getId());
+    }
+
+    @OnMessage
+    public String onStringMessage(String str, Session session) {
+        log.info("Test str is : " + str);
+        return str;
+    }
+
+    @OnClose
+    public void onClose(String reasonText, int statusCode) {
+    }
+}
